backend/requirements.txt
fastapi
uvicorn[standard]
openai
pinecone
numpy
pandas
psycopg2-binary
sqlalchemy
alembic
python-dotenv
pydantic
pydantic-settings
httpx
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
black
flake8
mypy

sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
scikit-learn==1.7.1


